rework settings router
